requirements.txt
# Core ML and Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
torch==2.0.1
tensorflow==2.13.0

# Feature Engineering
scipy==1.11.1
numba==0.57.1
cython==0.29.35

# Network Analysis
scapy==2.5.0
pyshark==0.6
dpkt==1.9.8
netaddr==0.8.0

# Streaming and Messaging
kafka-python==2.0.2
pika==1.3.2
redis==4.6.0

# Web Framework and API
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.0.3
gunicorn==21.2.0

# Database and Storage
elasticsearch==8.8.0
pymongo==4.4.1
psycopg2-binary==2.9.7
sqlalchemy==2.0.19
alembic==1.11.1

# Time Series and Monitoring
prometheus-client==0.17.1
grafana-api==1.0.3
influxdb-client==1.36.1

# Visualization and Dashboards
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1

# MLOps and Experiment Tracking
mlflow==2.5.0
wandb==0.15.8
dvc==3.0.0
optuna==3.2.0

# Model Serving and Optimization
onnx==1.14.0
onnxruntime==1.15.1
torchserve==0.7.1

# Security and Cryptography
cryptography==41.0.3
pycryptodome==3.18.0
hashlib-compat==1.0.1

# Utilities
click==8.1.4
tqdm==4.65.0
joblib==1.3.1
pyyaml==6.0.1
python-dotenv==1.0.0
schedule==1.2.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-asyncio==0.21.1

# Development
black==23.7.0
flake8==6.0.0
mypy==1.4.1
pre-commit==3.3.3

# Jupyter and Notebooks
jupyter==1.0.0
ipykernel==6.25.0
nbconvert==7.6.0

# System and Process Management
psutil==5.9.5
py-cpuinfo==9.0.0
memory-profiler==0.61.0

# Network Security Tools
suricata-update==1.2.3
zeek==5.0.0
